Implement VSCode task for openapi generation
Adds a Visual Studio Code task to generate the openapi.json file using the swagger tool command
For that to work a dependency on the `Swashbuckle.AspNetCore.Swagger` package has been added to allow that assembly to be run the `dotnet swagger` command.

A lean Host wrapper has been added to the `Main()` function to initialize the Jellyfin Server to a point where swagger can complete the openapi generation. That host wrapper initializes the JellyfinApi and JellyinfApiSwagger services.

Furthermore expanded the build task to support both Release and Debug configurations.
Finally added a launch task that can launch the jellyfin server without the need to run through the debugger.

resolves #12347