Prevent server from starting if the ffmpeg path is invalid
We've received numerous reports from users that Jellyfin is failing to play certain files, and it turns out that many of these issues are due to invalid FFmpeg configurations, such as incorrect installation, stale configuration values, or an unsupported FFmpeg version. The server currently still starts in this situation, leading users to mistakenly believe it's a server bug rather than a configuration error.

This PR makes the server be prevented from starting and print a helper text to help the user fix the installation if the FFmpeg path is not found, not supported, or not executable.

Signed-off-by: gnattu <gnattuoc@me.com>