Add MediaStream.ReferenceFrameRate for problematic video files
FFmpeg cannot reliably detect the average framerate of certain files and may report an unrealistic 1000fps. However, in many such cases, the RealFrameRate field remains accurate and reflects the actual playback framerate. Add a new ReferenceFrameRate field for use cases where AverageFrameRate is preferred but allows a fallback to RealFrameRate when necessary. This PR has replaced all current usages that rely on AverageFrameRate but also does a fallback with this new field.

Signed-off-by: gnattu <gnattuoc@me.com>