Never set RemoveOldMetadata when replacing all metadata.
When RemoveOldMetadata is set it will clear all existing metadata from items before applying the new information. This breaks virtual seasons as episodes will never have a ParentIndexNumber set unless the season is part of the episode file name (which in a lot of cases, it is not). The fallback to "season 1" is only available in EpisodeResolver which is not called during metadata refresh.
When the ParentIndexId is reset to null the RemoveObsoleteSeasons method in SeriesMetadataService will remove the virtual season(s) as it thinks there are no episodes in them.