Add perf tradeoff mode to image extractor
On low-end CPUs, image extraction can be quite slow because the CPU has to decode and analyze 24 frames to pick the best one. If the input is an HDR image, it also needs to perform tone mapping. On a Raspberry Pi 4 level of performance, this usually means more than 10 seconds per image extracted, and for 4K HDR inputs, it may even timeout. This PR adds a tradeoff mode where the extracted image is the single frame that was picked. Ideally it would come from one of the key frame which is very fast to extract, but if that failed it will fallback to extract the slower normal frame. The frame picked by this mode is worse and less representative, but the process is significantly faster.