Infer more audio codec from containers
This is mainly for some broken legacy clients that will unlikely to receive proper fix for audio transcoding profile and migrate to our new universal audio controller anytime soon. These legacy clients used an outdated Emby-era audio transcoding pipeline, where the client neither explicitly specifies the desired transcoding formats nor properly indicates the supported audio codecs in its request. Instead, they rely on just providing a container name in the request URL, hoping the server can guess the correct audio codec. However, this guessing is often inaccurate as the container-to-codec mapping has been hard-coded and outdated for six years. When the server cannot guess it will just return "copy" which will obviously fail because the client is requesting transcoding. This PR makes the infer logic more robust to only infer from containers that does not have the same name as the audio codec, and pick the "should be compatible" codec for the container. This may still fail in edge cases if the client is weird enough, but should cover most use cases.