Add a small tolerance value to remux fps check
Some videos have a recorded FPS that is slightly higher than the intended FPS. If this intended FPS is the maximum value requested by the device, it can cause the server to trigger transcoding. However, such minor statistical error do not affect remux playback, add a slight tolerance to enable remux for this case.