Fix trickplay extraction ffmpeg error-handling
If ffmpeg runs to completion, i.e. isn't killed due to being
unresponsive, its exit code is >0 in case of an error. However,
the current logic only throws an exception in case the exit code
is -1, which is the virtual exit code assigned in case ffmpeg has
been killed prematurely.

This commit simplifies the logic to throw an exception if ffmpeg
didn't run to completion (as before) or exited with a non-zero
exit code (new).