Normalizer cleanup
In PR 14648 new analysis for isAnamorphic was added. The string comparisons there were added as StringComparison.OrdinalIgnoreCase, but the actual strings are always digits:digits (e.g. 3:4) so there's no alphabetic values to ignore case of... since StringComparision.Ordinal is much faster, switch to that.
Also, morphed all the private methods that should be to static since that eliminates a null-ref check on the this (at call-sites) and also makes the JIT happy.