feat: add option to require file size change for media refreshes
adds a new system setting, `FileChangeRequireSizeChange`, which can be
configured in `system.xml`. When set to `true`, the server will only
consider a media file to have changed if both its modification time and
file size have changed.

This change is based on the discussions in #13839 and makes
the behaviour reintroduced in #14674 configurable. The default value
is `false`, preserving the default behaviour.