Make priority class setting more robust
Currently the tasks simply crashes because the user running jellyfin server does not have the permission to change priority of a process. The below normal is a best-effort optimization and should not crash the task entirely. Add a try catch to make the task to move on even when priority cannot be set.