Enforce more strict webm check
The webm spec only supports video and video streams, and other type of streams in the container should disqualify its webm container type and be treated as mkv.

Some tools might allow users to add subtitle streams to webm and still claim it is webm, but such files are not guranteed to play directly everywhere, and the actuall support is implementation specific.

With the advent of the (unreliable) native mkv support in firefox, this will be required to properly trigger remux for such files.